split heroic module into heroes and villains modules. fixed oruting too.
