+++ src/client/app/heroes/hero-detail.component.scss
