+++ src/client/app/heroes/hero-list.component.scss
