+++ src/client/app/heroes/heroes-routing.module.ts
@@ -1,0 +1,12 @@
+import { NgModule } from '@angular/core';
+import { Routes, RouterModule } from '@angular/router';
+
+import { HeroListComponent } from './hero-list.component';
+
+const routes: Routes = [{ path: '', pathMatch: 'full', component: HeroListComponent }];
+
+@NgModule({
+  imports: [RouterModule.forChild(routes)],
+  exports: [RouterModule]
+})
+export class HeroesRoutingModule {}
