+++ src/client/app/heroes/heroes.module.ts
@@ -1,0 +1,14 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { ReactiveFormsModule } from '@angular/forms';
+
+import { HeroesRoutingModule } from './heroes-routing.module';
+import { HeroDetailComponent } from './hero-detail.component';
+import { HeroListComponent } from './hero-list.component';
+
+@NgModule({
+  imports: [CommonModule, ReactiveFormsModule, HeroesRoutingModule],
+  exports: [HeroListComponent, HeroDetailComponent],
+  declarations: [HeroListComponent, HeroDetailComponent]
+})
+export class HeroesModule {}
